requirements.txt
# Top-level requirements for Streamlit Currency Detector
# Based on the working environment from money_env
# This minimal set of packages helps avoid dependency conflicts.
# Other packages like torch, numpy, and opencv-python will be installed
# automatically as dependencies of the packages below.

streamlit==1.44.1
ultralytics==8.3.108
tensorflow==2.17.0
keras==3.7.0
numpy==1.26.4
Pillow==11.2.1
